configure: Standardize string test with 'x' prefix
Either we do it or we don't. It is still good practice to do it because
apparently certain old shell have trouble comparing empty strings. Also, if
the variable is expanded with something that contains a dash, it will be
passed to "test" instead of evaluated by it.

Furthermore, the lack of 'x' in the tested string made some features not to
work and one of them is building OpenSSL statically.

Part of #33624

Signed-off-by: David Goulet <dgoulet@torproject.org>