openssl: Remove uneeded ifndef on OPENSSL_VERSION
The OPENSSL_VERSION is a define in the OpenSSL library that indicate a type of
version that can be passed to OpenSSL_version() in order to get the string
representation of the OpenSSL version.

OPENSSL_VERSION is defined in <openssl/crypto.h> which is includes by almost
all headers. Furthermore, SSLEAY_VERSION only exists for API compatibility and
defaults to OPENSSL_VERSION.

When statically building OpenSSL, we end up redefining OPENSSL_VERSION which
creates a compilation error because we include our compat OpenSSL header
_before_ any OpenSSL headers.

Closes #33624

Signed-off-by: David Goulet <dgoulet@torproject.org>