Remove tor.sh from .gitignore
The files contrib/dist/tor.sh and contrib/dist/suse/tor.sh should not
exist anymore, but when building in same repository for older versions
of Tor, these files may return and not be cleaned up properly. This
commit unignores those files from .gitignore, in order to make it clear
that these files are no longer auto-generated and should be cleaned up.