Only treat canonical-to-peer channels as indicating reachability.
Previously we'd treat an incoming circuit as indicating orport
reachability, even if it was to an ORPort other than the one we were
trying to test.  Now we only treat it as indicating reachability
when it is on a canonical-to-peer channel: one where the address
matched our published address when the channel was made.

Fixes bug introduced in 4f5192b2803c706 in 0.1.0.1-rc, back when we
added self-reachability tests. Closes 20165.