relay: Query our cache when deciding for dummy descriptor fetch
Instead of looking at the "Address" option alone, instead check if we have an
address in our cache (that is discovered by tor). If not, then it tells us
that tor does not have an address to work with so we can then ask a directory
authority for a suggestion.

Related #2178

Signed-off-by: David Goulet <dgoulet@torproject.org>