hs: Collect upload descriprot attempts and success metrics
This is a bit of a more complex commit. In order to track properly, on a per
descriptor level, the attempt and success of an upload, we have to keep a
counter in the descriptor object.

The metrics data concept is that it tracks success and attempts as a gauge so
if there is any discrepancy between the two, we can infer availbaility issues.

So those new counters in the descriptor are used to _decrement_ the per
service metrics data when attempting to re-upload the descriptor. In effect,
they put the metrics back to 0 for that specific descriptor so the next upload
attempt can be tracked properly.

Part of #40063

Signed-off-by: David Goulet <dgoulet@torproject.org>