config: Catch missing Bridge for ClientTransportPlugin
First of all, tor doesn't support multiple transport name on a single
*TransportPlugin line so get rid of that. The manpage has never documented
that feature and it complexifies the line for no reason. A user can just use
multiple ClientTransportPlugin line instead.

Second, if no Bridge line is found, error immediately and exit. Don't leave
the Bridge line alone there leading to bugs like #25528

Third, when making sure we have a Bridge line with a ClientTransportPlugin, we
now check at the managed proxy list. In order to pull this off, the transport
name had to be added to the managed_proxy_t object so we can match for which
transport it is used for. And thus justifying the removal of multiple
transport support which adds no benefit except complexity.

Fixes #40106

Signed-off-by: David Goulet <dgoulet@torproject.org>